backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Google Cloud / Vertex AI
google-cloud-aiplatform==1.38.1
google-auth==2.27.0

# ElevenLabs
elevenlabs==1.1.2

# Kafka / Confluent
confluent-kafka==2.3.0

# Database
asyncpg==0.29.0
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Logging and monitoring
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
